requirements.txt
abcmeta==2.0.0
Django==4.0.4
djangorestframework==3.13.1
